+++ NEWS.md
@@ -9,6 +9,7 @@
 * New `getReactableState()` function to get the state of a reactable instance in Shiny.
 * In `reactable()`, the `rowClass`, `rowStyle`, and `details` JavaScript functions
   now receive a `rowInfo.selected` property indicating whether the row is selected.
+* `reactable()` gains a `theme` argument to customize the default styling of a table.
 * `reactable()` gains a `language` argument to customize the language strings in a table.
 * `colDef()` gains a `"median"` aggregate function to calculate the median of numbers.
 * The row selection column can now be customized using `".selection"` as the column name.